+++ crates/alembic-cli/examples/provisioning_emitter_adapter.rs
@@ -1,8 +1,9 @@
-use alembic_core::{Schema, TypeName};
-use alembic_engine::{
-    alembic_external_main, ApplyReport, ExternalAdapter, ExternalCapabilities, ExternalObject,
-    ExternalRole, Op, ProvisionReport, StateData,
+use alembic_adapter_sdk::alembic_external_main;
+use alembic_adapter_sdk::{
+    ApplyReport, ExternalAdapter, ExternalCapabilities, ExternalObject, ExternalRole, Op,
+    ProvisionReport, StateData,
 };
+use alembic_core::{Schema, TypeName};
 use anyhow::Result;
 
 alembic_external_main!(ProvisioningEmitterAdapter::default());
@@ -13,6 +14,8 @@ alembic_external_main!(ProvisioningEmitterAdapter::default());
 pub struct ProvisioningEmitterAdapter {}
 
 impl ExternalAdapter for ProvisioningEmitterAdapter {
+    type Error = anyhow::Error;
+
     fn setup(&mut self, _configuration: &serde_yaml::Value) -> Result<()> {
         Ok(())
     }